Change to verification from messy max-len to limitSize

+++ lib/blockchain/tx/data/gas.dart
@@ -4,14 +4,14 @@ import 'package:protolith/blockchain/tx/transaction.dart';
 
 mixin GasTxData<M extends BlockMeta> on Transaction<M> {
 
-  int _gas;
+  BigInt _gas;
   /// QUANTITY - gas provided by the sender.
-  int get gas => _gas;
-  set gas(int v) => _gas = hashed(v);
+  BigInt get gas => _gas;
+  set gas(BigInt v) => _gas = hashed(v);
 
-  int _gasPrice;
+  BigInt _gasPrice;
   /// QUANTITY - gas price provided by the sender in Wei.
-  int get gasPrice => _gasPrice;
-  set gasPrice(int price) => _gasPrice = hashed(price);
+  BigInt get gasPrice => _gasPrice;
+  set gasPrice(BigInt price) => _gasPrice = hashed(price);
 
 }
